# Conflicts:
#	kairos-planner/optaplanner-vrp-taskplanning/src/main/resources/solution.xml

+++ dto/src/main/java/com/kairos/response/dto/web/wta/ConsecutiveRestPartOfDayWTATemplateDTO.java
@@ -24,7 +24,7 @@ public class ConsecutiveRestPartOfDayWTATemplateDTO extends WTABaseRuleTemplateD
     private long daysWorked;
     protected List<PartOfDay> partOfDays = new ArrayList<>();
 
-    private List<Long> plannedTimeIds = new ArrayList<>();
+    private List<BigInteger> plannedTimeIds = new ArrayList<>();
     private List<BigInteger> timeTypeIds = new ArrayList<>();
     protected float recommendedValue;
     private MinMaxSetting minMaxSetting;
@@ -39,11 +39,11 @@ public class ConsecutiveRestPartOfDayWTATemplateDTO extends WTABaseRuleTemplateD
     }
 
 
-    public List<Long> getPlannedTimeIds() {
+    public List<BigInteger> getPlannedTimeIds() {
         return plannedTimeIds;
     }
 
-    public void setPlannedTimeIds(List<Long> plannedTimeIds) {
+    public void setPlannedTimeIds(List<BigInteger> plannedTimeIds) {
         this.plannedTimeIds = plannedTimeIds;
     }
